# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClinicalCompass API",
    "summary": "Authentication API with database-backed opaque tokens.",
    "description": "\n**ClinicalCompass API** \u2014 username/password authentication using opaque,\nserver-side tokens (no JWT).\n\n### Auth flow\n1. `POST /auth/register` to create a user.\n2. `POST /auth/login` with the username + password to receive an opaque\n   `access_token`.\n3. Send the token as `Authorization: Bearer <access_token>` on protected\n   endpoints.\n4. `POST /auth/logout` to revoke the token.\n\nTokens are random strings stored in the database and expire after 24 hours.\n",
    "contact": {
      "name": "ClinicalCompass",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/specialities/{speciality_id}/modules": {
      "get": {
        "tags": [
          "specialities"
        ],
        "summary": "List modules in a speciality",
        "operationId": "list_modules_specialities__speciality_id__modules_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "speciality_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Speciality Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleOut"
                  },
                  "title": "Response List Modules Specialities  Speciality Id  Modules Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Speciality not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/modules": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "List/search modules (paginated)",
        "description": "Returns a page of modules. Filter by `specialities` (one or more speciality ids, repeatable) and/or `search_query` (case-insensitive title match). Paginate with `limit`/`offset`; `next`/`previous` are URLs.",
        "operationId": "list_modules_modules_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "specialities",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to these speciality ids (repeatable)",
              "title": "Specialities"
            },
            "description": "Filter to these speciality ids (repeatable)"
          },
          {
            "name": "search_query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive match on title or description",
              "title": "Search Query"
            },
            "description": "Case-insensitive match on title or description"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max items to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModulePage"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/modules/count": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "Count of modules",
        "description": "Returns the total number of modules across all specialities.",
        "operationId": "count_modules_modules_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/modules/{module_id}": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "Get a module (including its form)",
        "operationId": "get_module_modules__module_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "module_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Module Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Module not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/modules/{module_id}/form": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "Get a module's form (stages -> fields)",
        "operationId": "get_module_form_modules__module_id__form_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "module_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Module Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleForm"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Module not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/modules/{module_id}/auth-guide": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "Get a module's authorization guide (CPT/ICD/payer reference)",
        "description": "Returns the module's prior-authorization guide (CPT codes, ICD-10 codes, payer rules, plus a verbatim `raw` block), or `null` when the module has no guide configured.",
        "operationId": "get_module_auth_guide_modules__module_id__auth_guide_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "module_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Module Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthGuideOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Module Auth Guide Modules  Module Id  Auth Guide Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Module not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthGuideOut": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "AuthGuideOut",
        "description": "A module's authorization guide, stored and returned verbatim.\n\nHybrid free-form shape: structured ``cptCodes``/``icd10Codes``/``payers``/\n``customSections`` for querying, plus a verbatim ``raw`` block (markdown +\ntables) that guarantees no source content is lost. Extra keys are allowed so\nheterogeneous guides are never rejected at the API boundary.",
        "example": {
          "cptCodes": [
            {
              "code": "50593",
              "description": "Ablation, renal tumor(s)..."
            }
          ],
          "icd10Codes": [
            {
              "code": "C64.1",
              "description": "Malignant neoplasm of right kidney"
            }
          ],
          "payers": [
            {
              "key": "cms",
              "label": "CMS / Medicare (LCD L35041)",
              "sections": []
            }
          ],
          "raw": {
            "markdown": "...",
            "tables": []
          },
          "sourceFile": "RenalCryoablationAuthGuide.tsx",
          "title": "Renal Cryoablation \u2014 Authorization Guide"
        }
      },
      "CountOut": {
        "properties": {
          "resource": {
            "type": "string",
            "title": "Resource",
            "examples": [
              "specialities"
            ]
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "examples": [
              42
            ]
          }
        },
        "type": "object",
        "required": [
          "resource",
          "count"
        ],
        "title": "CountOut",
        "description": "A labeled count of a resource."
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Invalid username or password"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Standard FastAPI error envelope."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModuleDetail": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "examples": [
              "Rectal Cancer"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "examples": [
              "Staging and treatment pathways"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "speciality_id": {
            "type": "integer",
            "title": "Speciality Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "/uploads/images/abc123.png"
            ]
          },
          "form": {
            "$ref": "#/components/schemas/ModuleForm"
          },
          "appeal_letter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appeal Letter"
          },
          "appeal_letter_template": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appeal Letter Template",
            "description": "Data-driven recipe for personalizing the appeal letter from a submission's answers (defaults, derived values, conditional sections, and a {{placeholder}} template)."
          },
          "auth_guide": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Guide",
            "description": "Prior-authorization guide: CPT codes, ICD-10 codes, payer-specific coverage/documentation rules, plus a verbatim `raw` block."
          }
        },
        "type": "object",
        "required": [
          "title",
          "id",
          "speciality_id"
        ],
        "title": "ModuleDetail",
        "description": "A single module including its form."
      },
      "ModuleForm": {
        "properties": {
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ModuleForm",
        "description": "A module's form, stored and returned verbatim.\n\nShape: {\"steps\": [{\"id\", \"shortLabel\", \"title\", \"fields\": [field, ...]}]}.\nEach field is free-form (id, name, label, type, required, defaultValue,\noptions, optionLayout, min, max, step, unit, placeholder, keyboardType,\nmultiline, ...), so the frontend owns the exact field schema.",
        "example": {
          "steps": [
            {
              "fields": [
                {
                  "defaultValue": "T3",
                  "id": "t-stage",
                  "label": "T Stage",
                  "name": "tStage",
                  "options": [
                    {
                      "label": "T1 \u2014 Submucosa",
                      "value": "T1"
                    },
                    {
                      "label": "T3 \u2014 Through muscularis propria",
                      "value": "T3"
                    }
                  ],
                  "required": true,
                  "type": "select"
                },
                {
                  "defaultValue": 0,
                  "id": "tumor-size",
                  "label": "Tumor Size",
                  "max": 15,
                  "min": 0,
                  "name": "tumorSize",
                  "step": 0.5,
                  "type": "slider",
                  "unit": "cm"
                }
              ],
              "id": "staging",
              "shortLabel": "Staging",
              "title": "Tumor Staging (MRI-based)"
            }
          ]
        }
      },
      "ModuleOut": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "examples": [
              "Rectal Cancer"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "examples": [
              "Staging and treatment pathways"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "speciality_id": {
            "type": "integer",
            "title": "Speciality Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "/uploads/images/abc123.png"
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "id",
          "speciality_id"
        ],
        "title": "ModuleOut"
      },
      "ModulePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModuleOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next",
            "description": "URL of the next page, or null on the last page",
            "examples": [
              "http://localhost/modules?limit=20&offset=20"
            ]
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous",
            "description": "URL of the previous page, or null on the first page",
            "examples": [
              null
            ]
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ModulePage",
        "description": "A page of modules with next/previous page links."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}